simplifying parallelized API
